requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.6
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
djangorestframework==3.14.0
drf-yasg==1.21.4
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
yarl==1.8.2
